Add ExercisesLLMService and GoalsLLMService for enhanced exercise and goal management
